Ticket #1291

+++ modules/boonex/timeline/install/sql/install.sql
@@ -257,7 +257,7 @@ CREATE TABLE IF NOT EXISTS `bx_timeline_reports_track` (
 CREATE TABLE IF NOT EXISTS `bx_timeline_hot_track` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `event_id` int(11) NOT NULL default '0',
-  `value` tinyint(4) NOT NULL default '0',
+  `value` int(11) NOT NULL default '0',
   PRIMARY KEY (`id`),
   UNIQUE KEY `event_id` (`event_id`)
 );
